Update index.md
Got stuck where my assembly wasn't reliably being picked up for subscription; tracked down the issue to be an assembly with version 0.0.1.1 apparently not being considered compliant with assembly with version 0.0.128.0.  If I force the pair to be 0.0.1.0 & 0.0.128.0, or 0.0.1.1 & 0.0.128.1, then we have success.  I can work with this for now, but do you agree this is a potential issue?