Update index.md
I added a single paragraph. But although it might be clear or correct what it states (I have you to verify that ;-) ) I'm not sure of readers of the document will understand.

I've thought about a code snippet of a single line, which seems ridiculous. Also because it's not NSB version related, it's .NET related.
I've also thought about putting a triple ` addition, like in the step-by-step page (http://docs.particular.net/samples/step-by-step/)
But I'm not sure if I'm allowed to use that.

So I'd like some feedback