MsmqToSqlRelay v6 sample
This sample relays events that are published from an MSMQ publisher to downstream SqlSubscribers using a SqlRelay.
It uses simple native bridging to move messages from a MSMQ to SQL table that the SqlRelay endpoint listens to.
The SqlRelay processes these events and publishes the events for its downstream SqlSubscribers.