Added glossary
Was discussing terms not being clear in a document and decided perhaps we could use a glossary. Reading up on terms like `consumer` and the difference between `logical endpoint` and `endpoint instance` is missing at the moment, whereas we are using them throughout the documentation. This makes it available and perhaps others/other-documents could make use of it as well. 

It's just a first kick off of this document to see if it's usable, before spending too much time on it. If we agree this could be usefull, we can extend it before merging it.