Explicit calculations of the queue length
The sent/received counters were increased to bigger values to show that they are not a part of the calculation of the queue length but only are used to derive `Messages in queue from this link`. The table below was enhanced with one table to provide calculations. All the values are unique, hence reader can link them between these two tables lexicographically.