Update doc to reference audit/error log queues
Fixes Particular/ServiceControl/#1568 where there was some confusion with respect to forwarding _to_ the log queues, versus the queues being called forwarding queues. The queues are forwarded _to_, they are not doing the forwarding themselves.

However, I also asked the question as to whether or not the SCMU UI should also be updated to better reflect the nature of the queue to be created? E.g. change _ERROR FORWARDING QUEUE NAME_ to _ERROR FORWARDING LOG QUEUE NAME_.