content verification
System.Exception : Some snippets exist that were not used in markdown. Or, the snippet may be used in some versions, but not all, in which case you need to remove the snippet from the version in which it's not needed. These are the unused snippets and the versions in which they weren't used: configuringinmemory (>= 7.2.0 && < 8.0.0)