Use common APM names with tagged metadata
We are using NServiceBus and DataDog and found that the original implementation of this integration created far too many metrics and was nearly impossible to create usable dashboards. What wound up really helping is using the common APM names for specific metrics, and then added things liked EndpointName and MessageType as tags on that metric. This allows you to aggregate all of a single metric for all messages combined (i.e. total cluster throughput), as well as break things down by individual endpoints, or to the fine grained message type.