Update index.md
Particular Team.  I had to read this too many times for everything to sink in (keep in mind, i might be slow :).  I was constantly asking my self how is it possible for the messages to be safely sent if that is occurring as part of the handler logic.  That is until I realized that there are two different processes occurring.  I realize the diagram in step 6 says that the transaction is committed, but I made the mistake of assuming there was still a parent transaction wrapped around the whole process (i.e. steps 1 through 10).  Subsequently, the diagram, may merit some updates as it does not make it clear that there are two distinct processes at play.  I also added some verbiage to high light that there are multiple process.  @seanfarmar